+++ react/__mocks__/vtex.product-summary-context.ts
@@ -1,0 +1,3 @@
+export const ProductSummaryContext = {
+  useProductSummary: jest.fn(),
+}