+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Updated the app version mentioned in README.md.
+
 ## [1.6.0] - 2025-07-16
 
 ### Added