+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Docs file for availabilityBadges
+- Tests file for availability Badges
+
 ## [1.6.1] - 2025-08-04
 
 ### Changed