+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.2] - 2025-09-23
+
 ### Removed
 
 - `vtex.react-portal` and `vtex.store-drawer` dependencies